+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Only show language chooser when more than one locale is available
-- Retrict items search date picker to collection's temporal extend.
+- Restrict Collection item search date picker to collection's temporal extent
+- Focus temporal extent filter for Collection item search on end of temporal extent
+- Disable temporal extent filter when a single date/time is provided as temporal extent in the Collection metadata
 
 ## [5.0.0-beta.1] - 2026-05-12